1. Agreement can be deleted only by the manager who created them

+++ src/main/java/lv/velexauto/velex/mvc/Protected/AgreementDeleteControler.java
@@ -1,8 +1,10 @@
 package lv.velexauto.velex.mvc.Protected;
 
-import lv.velexauto.velex.HelperClasses.DataValidateAssistant;
+import lv.velexauto.velex.HelperClasses.SecurityAssistant;
 import lv.velexauto.velex.database.AgreementDAO;
 import lv.velexauto.velex.database.DBException;
+import lv.velexauto.velex.domain.Agreement;
+import lv.velexauto.velex.domain.Employee;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
@@ -22,8 +24,18 @@ public class AgreementDeleteControler {
     @Qualifier("HibernateDAOAgreement")
     private AgreementDAO agreementDAO;
 
+    @Autowired
+    @Qualifier("SecurityAssistant")
+    private SecurityAssistant securityAssistant;
+
     @RequestMapping(value = {"protected/deleteAgreement"}, method = RequestMethod.POST)
-    public @ResponseBody String deleteAgreement(@RequestParam(value = "agreementID") long agreementID) {
+    public @ResponseBody String deleteAgreement(@RequestParam(value = "agreementID") long agreementID) throws DBException {
+
+        long currentManagerID = securityAssistant.getCurrentEmployee().getEmployeeId();
+        long agreementManagerID = agreementDAO.getById(agreementID).getEmployee().getEmployeeId();
+
+        if (currentManagerID != agreementManagerID) return "can't delete";
+
         try {
             agreementDAO.delete(agreementID);
             return "success";
